refactor(host-selfhost): extract SSO admission helpers, trim comments
